Renamed examples and streamlined interface to run experiments as a module
