+++ examples/mnist_lenet5.py
@@ -1,6 +1,4 @@
-from collections.abc import Callable
 import torch
-from torch import Tensor
 from torch import nn
 from torchvision import transforms, datasets
 from torch.utils.data import DataLoader
@@ -8,36 +6,13 @@ from torch.utils.data import DataLoader
 from tqdm import tqdm
 
 from bitnet.nn.bitlinear import BitLinear
-from seed import set_seed
+from bitnet.nn.bitconv2d import BitConv2d
+from bitnet.models.lenet5 import LeNet
+from bitnet.metrics import Metrics
+from bitnet.seed import set_seed
 
 
 device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-
-
-class Net(nn.Module):
-    def __init__(
-            self,
-            linear_layer: Callable,
-            input_size: int,
-            hidden_size: int,
-            num_classes: int
-        ) -> None:
-        super(Net, self).__init__()
-        self.linear_layer = linear_layer
-        self.fc1 = linear_layer(input_size, hidden_size)
-        self.fc2 = linear_layer(hidden_size, num_classes)
-
-
-    @property
-    def __name__(self) -> str:
-        return "BitNet" if self.linear_layer == BitLinear else "FloatNet"
-
-
-    def forward(self, _input: Tensor) -> Tensor:
-        out = _input.flatten(start_dim=1)
-        out = torch.relu(self.fc1(out))
-        out = self.fc2(out)
-        return out
 
 
 def train_model(
@@ -66,7 +41,8 @@ def train_model(
         pbar.close()
 
 
-def test_model(model: nn.Module, test_loader: DataLoader):
+def test_model(model: nn.Module, test_loader: DataLoader) -> tuple[dict[str, float], Metrics]:
+    metrics_used: Metrics = Metrics.ACCURACY
     model.eval()
     correct: int = 0
     total: int = 0
@@ -77,17 +53,15 @@ def test_model(model: nn.Module, test_loader: DataLoader):
             _, predicted = torch.max(outputs.data, 1)
             total += labels.size(0)
             correct += (predicted == labels).sum().item()
-    accuracy = 100 * correct / total
-    print(f'Accuracy of {model.__name__}: {accuracy:.2f}%')
-    return accuracy
+    metric: float = 100 * correct / total
+    return {model.__name__: metric}, metrics_used
 
 
-def main():
+def run(seed: int | None) -> tuple[dict[str, float], Metrics]:
 
-    set_seed()
+    set_seed(seed)
+    return_value: dict[str, float] = {}
 
-    input_size: int         = 784
-    hidden_size: int        = 100
     num_classes: int        = 10
     learning_rate: float    = 1e-3
     num_epochs: int         = 5
@@ -99,8 +73,8 @@ def main():
     ])
 
     print(f"Testing on {device=}")
-    bitnet = Net(BitLinear, input_size, hidden_size, num_classes).to(device)
-    floatnet = Net(nn.Linear, input_size, hidden_size, num_classes).to(device)
+    bitnet = LeNet(BitLinear, BitConv2d, num_classes, 1, 28).to(device)
+    floatnet = LeNet(nn.Linear, nn.Conv2d, num_classes, 1, 28).to(device)
 
     bitnet_optimizer = torch.optim.Adam(bitnet.parameters(), lr=learning_rate)
     floatnet_optimizer = torch.optim.Adam(floatnet.parameters(), lr=learning_rate)
@@ -110,22 +84,24 @@ def main():
     train_dataset = datasets.MNIST('./mnist_data', train=True, download=True, transform=transform)
     test_dataset = datasets.MNIST('./mnist_data', train=False, download=True, transform=transform)
 
-    set_seed()
+    set_seed(seed)
     train_loader = DataLoader(train_dataset, batch_size=batch_size, shuffle=True)
     test_loader = DataLoader(test_dataset, batch_size=batch_size, shuffle=False)
     train_model(bitnet, train_loader, bitnet_optimizer, criterion, num_epochs)
-    bitnet_accuracy: float = test_model(bitnet, test_loader)
+    test_model(bitnet, test_loader)
+    results, metrics_used = test_model(bitnet, test_loader)
+    return_value.update(results)
 
-    set_seed()
+    set_seed(seed)
     train_loader = DataLoader(train_dataset, batch_size=batch_size, shuffle=True)
     test_loader = DataLoader(test_dataset, batch_size=batch_size, shuffle=False)
     train_model(floatnet, train_loader, floatnet_optimizer, criterion, num_epochs)
-    floatnet_accuracy: float = test_model(floatnet, test_loader)
+    test_model(floatnet, test_loader)
+    results, metrics_used = test_model(floatnet, test_loader)
+    return_value.update(results)
 
-    difference = abs(floatnet_accuracy - bitnet_accuracy)
-    diff_threshold = 1.0
-    assert difference < diff_threshold, "Accuracy must be within 1%"
+    return return_value, metrics_used
 
 
 if __name__ == "__main__":
-    main()
+    print(run(None))
