--- examples/seed.py
@@ -1,16 +1,0 @@
-import random
-
-import numpy as np
-import torch
-
-
-def set_seed(seed_value: int = 53):
-    """Sets the seed for reproducibility."""
-    torch.manual_seed(seed_value)
-    torch.cuda.manual_seed(seed_value)
-    torch.cuda.manual_seed_all(seed_value)
-    np.random.seed(seed_value)
-    random.seed(seed_value)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
-
